Trying to solve some Win/Linux compatibility issues.

+++ cRio_Daq_cpp/trunk/cRio_DAQ/src/RealTimer.h
@@ -23,7 +23,7 @@ public:
 	RealTimer();
 	virtual ~RealTimer();
 	void start();
-	float stop();
+	double stop();
 	float getResolution();
 private:
 	struct timeval startTime;
